Split the VS PR stage separately from the upload steps
There is sometimes an issue with the clone/create PR step. When that
happens, the stage isn't requeuable because the VSTS upload step fails
when its run again because the payload is already there. This update
allows the step to clone the VS repo and create the PR to be rerun due
to issues independent of the Upload step.